+++ src/bldc.c
@@ -70,7 +70,7 @@ uint32_t buzzerPattern = 0;
 uint32_t buzzerTimer    = 0;
 
 uint8_t enable = 0;
-volatile long long bldc_counter = 0;
+volatile int64_t bldc_counter = 0;
 volatile unsigned  bldc_count = 0;
 
 volatile unsigned  bldc_count_per_hall_counter[2] = {0,0};
@@ -165,7 +165,7 @@ void DMA1_Channel1_IRQHandler() {
   unsigned int hall[2];
   // note: the h_timer_hall IS a 100khz clock, which we read here just to have a high accuracy time
   // updated at 16khz.
-  unsigned long time_in = DWT->CYCCNT;
+  uint32_t time_in = DWT->CYCCNT;
 
   __disable_irq(); // but we want both values at the same time, without interferance
   hall[0] = (~(LEFT_HALL_U_PORT->IDR & (LEFT_HALL_U_PIN | LEFT_HALL_V_PIN | LEFT_HALL_W_PIN))/LEFT_HALL_U_PIN) & 7;
@@ -347,7 +347,7 @@ void DMA1_Channel1_IRQHandler() {
   }
 
   /* do something */
-  unsigned long time_out = DWT->CYCCNT;
+  uint32_t time_out = DWT->CYCCNT;
   if (time_out < time_in) {
     time_out += 0x80000000;
     time_in += 0x80000000;
